SSR checkout data: fetch cart, countries, addresses server-side
Split the checkout page into a Server Component (page.tsx) that fetches
all initial data during SSR, and a Client Component (CheckoutPageContent)
that receives it as props.

Before: page rendered a loading skeleton, hydrated, fired 5 server action
calls from the client, then showed the checkout UI.

After: Server Component fetches cart, market, countries, addresses, and
auth status in parallel during the initial request. The client component
initializes with real data — no loading skeleton, no waterfall.

loadOrder() is kept as a refresh function for coupon changes, express
checkout completion, etc.
